pyproject.toml
[project]
name = "articlehub-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "fastapi[all]>=0.116.1",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "pymongo[srv]>=4.14.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=7.0.0",
    "requests>=2.32.5",
]
lint = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
]

[tool.black]
line-length = 119
exclude = '''
(
  /(
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
)
'''

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = '.venv'

[tool.isort]
line_length = 119

[tool.pytest.ini_options]
addopts = "--cov --cov-report term-missing"
